requirements.txt
python-telegram-bot==13.7
requests==2.31.0
 google-cloud-logging==3.8.0